code/gen-ai/mcp/todo-mcp-server/requirements.txt
fastmcp>=0.1.0
snowflake-id>=1.0.0
